Bug fixes and adjustments: - In randomizer, Fixed a bug with randomboost 0: it needed to be supplying None to set_randomness_multiplier when randomboost is set to 0. - In beyondchaos, moved the custom NonDaemonThread and NonDaemonPool into a separate file, customthreadpool, to work around a python bug related to having the multiprocessing code within the same module that uses it. - In beyondchaos.textChanged, separated QSpinBox and QDoubleSpinBox to fix an error about depreciating implicit int conversions.
